initial demo for patching configs capability
