+++ examples/tutorials/config_sample_2pop_outcomes.yml
@@ -1,50 +1,3 @@
-name: sample_2pop
-setup_name: minimal
-start_date: 2020-02-01
-end_date: 2020-05-31
-nslots: 1
-
-subpop_setup:
-  geodata: model_input/geodata_sample_2pop.csv
-  mobility: model_input/mobility_sample_2pop.csv
-
-initial_conditions:
-  method: SetInitialConditions
-  initial_conditions_file: model_input/ic_2pop.csv
-  allow_missing_subpops: TRUE
-  allow_missing_compartments: TRUE
-
-compartments:
-  infection_stage: ["S", "E", "I", "R"]
-
-seir:
-  integration:
-    method: rk4
-    dt: 1
-  parameters:
-    sigma:
-      value: 1 / 4
-    gamma:
-      value: 1 / 5
-    Ro:
-      value: 2.5
-  transitions:
-    - source: ["S"]
-      destination: ["E"]
-      rate: ["Ro * gamma"]
-      proportional_to: [["S"],["I"]]
-      proportion_exponent: ["1","1"]
-    - source: ["E"]
-      destination: ["I"]
-      rate: ["sigma"]
-      proportional_to: ["E"]
-      proportion_exponent: ["1"]
-    - source: ["I"]
-      destination: ["R"]
-      rate: ["gamma"]
-      proportional_to: ["I"]
-      proportion_exponent: ["1"]
-
 outcomes:
   method: delayframe
   outcomes:
