+++ flepimop/gempyor_pkg/setup.cfg
@@ -35,6 +35,7 @@ install_requires =
     dask
     scipy
     graphviz
+    pydantic
 
 # see https://stackoverflow.com/questions/58826164/dependencies-requirements-for-setting-up-testing-and-installing-a-python-lib
 # installed for pip install -e ".[test]"
@@ -53,6 +54,7 @@ console_scripts =
     gempyor-seir = gempyor.simulate_seir:simulate
     gempyor-simulate = gempyor.simulate:simulate
     flepimop-calibrate = gempyor.calibrate:calibrate
+    flepimop-validate = gempyor.config_validator:validate
 
 [options.packages.find]
 where = src
