+++ flepimop/gempyor_pkg/src/gempyor/config_validator.py
@@ -5,6 +5,9 @@ from typing import Dict, List, Union, Literal, Optional, Annotated, Any
 from functools import partial
 from gempyor import compartments
 
+def validate(file_path: str):
+    return True
+
 def read_yaml(file_path: str) -> dict:
     with open(file_path, 'r') as stream:
         config = yaml.safe_load(stream)
@@ -14,11 +17,6 @@ def read_yaml(file_path: str) -> dict:
 def allowed_values(v, values):
     assert v in values
     return v
-
-# def parse_value(cls, values):
-#     value = values.get('value')
-#     parsed_val = compartments.Compartments.parse_parameter_strings_to_numpy_arrays_v2(value)
-#     return parsed_val
     
 class SubpopSetupConfig(BaseModel):
     geodata: str
